+++ src/assetpreviews/AudioPreview.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @copyright Copyright (c) vnali
  */
@@ -7,7 +8,6 @@ namespace vnali\studio\assetpreviews;
 
 use Craft;
 use craft\base\AssetPreviewHandler;
-
 use yii\base\NotSupportedException;
 
 /**
@@ -21,12 +21,17 @@ class AudioPreview extends AssetPreviewHandler
     public function getPreviewHtml(array $variables = []): string
     {
         $url = $this->asset->getUrl();
-
         if ($url === null) {
             throw new NotSupportedException('Preview not supported.');
         }
 
-        return Craft::$app->getView()->renderTemplate('studio/assets/_previews/audio',
+        $view = Craft::$app->view;
+        $view->startJsBuffer();
+        $variables['createField'] = $view->renderTemplate('studio/assets/_previews/chapter.twig');
+        $variables['createFieldJs'] = $view->clearJsBuffer(false);
+
+        return Craft::$app->getView()->renderTemplate(
+            'studio/assets/_previews/audio',
             array_merge([
                 'asset' => $this->asset,
                 'url' => $url,